Refactor utils to services
