+++ cli/modules/configuration/update.js
@@ -2,12 +2,13 @@ require('colors')
 const inquirer = require('inquirer')
 const Path = require('path')
 const Fs = require('fs')
-const hq = require('../../src')
-const { saveSettings, loadJson, saveJson } = require('../services/config')
-const { inspect } = require('../utils')
-const { makeChoices } = require('../utils/inquirer')
-const { indent, makeJson } = require('../utils/text')
-const { checkPath, checkPaths } = require('../services/common')
+const hq = require('../../../src')
+const { checkPath, checkPaths } = require('../../services/paths')
+const { saveSettings } = require('../../services/config')
+const { loadJson, saveJson } = require('../../utils/file')
+const { indent, makeJson } = require('../../utils/text')
+const { makeChoices } = require('../../utils/prompts')
+const { inspect } = require('../../utils')
 
 // ---------------------------------------------------------------------------------------------------------------------
 // helpers
