+++ cli/modules/integrations/debug.js
@@ -1,7 +1,7 @@
 const inquirer = require('inquirer')
-const hq = require('../../src')
-const { getPlugins } = require('../services/config')
-const { indent, makeJson } = require('../utils/text')
+const hq = require('../../../src')
+const { getPlugins } = require('../../services/config')
+const { indent, makeJson } = require('../../utils/text')
 
 // ---------------------------------------------------------------------------------------------------------------------
 // helpers
