+++ cli/modules/source/transformer/paths.js
