+++ cli/services/config.js
@@ -1,7 +1,16 @@
 const Path = require ('path')
-const Fs = require ('fs')
 const hq = require('../../src')
-const { compactJson } = require('./text')
+const { loadJson, saveJson } = require('../utils/file')
+const { indent, makeJson} = require('../utils/text')
+
+/**
+ * Log the current config to the terminal
+ */
+function showConfig () {
+  hq.load()
+  console.log()
+  console.log(indent(makeJson(hq.config)))
+}
 
 /**
  * Returns all plugins as a 2D hash of plugins and their options
@@ -96,51 +105,6 @@ function numAliases (load = false) {
   return Object.keys(hq.config.paths).length
 }
 
-function loadJson (filename, asText = false) {
-  const path = Path.resolve(filename)
-  try {
-    const text = Fs.readFileSync(path, 'utf8')
-    return asText
-      ? text
-      : JSON.parse(text)
-  }
-  catch (err) {
-    console.warn(`Could not load "${filename}"`)
-    return asText ? '' : null
-  }
-}
-
-function saveJson (filename, data, compact = false) {
-  // path
-  const path = Path.resolve(filename)
-
-  // get spacing
-  const text = loadJson(filename, true) || ''
-  const match = text.match(/^(\s+)"\w/)
-  const spacing = match ? match[1] : '  '
-
-  // get text
-  let json = JSON.stringify(data, null, spacing)
-
-  // compact arrays
-  if (compact) {
-    json = compactJson(json)
-  }
-
-  // save
-  return saveText(path, json)
-}
-
-function saveText (path, text) {
-  try {
-    return Fs.writeFileSync(path, text, 'utf8')
-  }
-  catch (err) {
-    console.warn(`Could not save "${Path.basename(text)}"`, err)
-    return null
-  }
-}
-
 function saveSettings (newSettings) {
   const data = loadJson('./package.json')
   if (data) {
@@ -157,6 +121,6 @@ module.exports = {
   getAliases,
   numAliases,
   saveSettings,
-  loadJson,
-  saveJson,
+  showConfig,
 }
+
