+++ tests/specs/cli/revert-source.spec.js
@@ -1,7 +1,7 @@
 const hq = require('../../../src')
 const { abs, inspect } = require('../../../src/utils')
-const { getAliases } = require('../../../cli/utils/config')
-const { toRelative } = require('../../../cli/source/paths')
+const { getAliases } = require('../../../cli/services/config')
+const { toRelative } = require('../../../cli/modules/source/transformer/paths')
 
 // ---------------------------------------------------------------------------------------------------------------------
 // helpers
