+++ tests/specs/cli/update-source.spec.js
@@ -1,7 +1,7 @@
 const hq = require('../../../src')
 const { abs, inspect } = require('../../../src/utils')
-const { getAliases } = require('../../../cli/utils/config')
-const { toAlias } = require('../../../cli/source/paths')
+const { getAliases } = require('../../../cli/services/config')
+const { toAlias } = require('../../../cli/modules/source/transformer/paths')
 
 // ---------------------------------------------------------------------------------------------------------------------
 // helpers
